backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
textblob==0.17.1
newsapi-python==0.2.7
python-multipart==0.0.6
httpx==0.25.2
asyncio
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10